Move KtorGoogleFormApi to common class of MultiPlatformProject
