+++ buildSrc/src/main/java/dependencies/Dep.kt
@@ -102,7 +102,9 @@ object Dep {
 
     object Ktor {
         val version = "1.0.1"
+        val clientCommon = "io.ktor:ktor-client-core:$version"
         val clientAndroid = "io.ktor:ktor-client-okhttp:$version"
+        val clientIos = "io.ktor:ktor-client-ios:$version"
         val jsonJvm = "io.ktor:ktor-client-json-jvm:$version"
     }
