+++ data/api-impl/build.gradle
@@ -1,7 +1,7 @@
 import dependencies.Dep
 
+apply plugin: 'kotlin-multiplatform'
 apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
 apply plugin: 'kotlinx-serialization'
 apply plugin: 'kotlin-kapt'
 
@@ -9,6 +9,24 @@ apply from: rootProject.file('gradle/android.gradle')
 android.libraryVariants.all {
     // TODO replace with https://issuetracker.google.com/issues/72050365 once released.
     it.generateBuildConfig.enabled = true
+}
+
+kotlin {
+    targets {
+        fromPreset(presets.android, 'android')
+
+        final def buildForDevice = project.findProperty("device")?.toBoolean() ?: false
+        final def iosPreset = (buildForDevice) ? presets.iosArm64 : presets.iosX64
+        fromPreset(iosPreset, 'ios') {
+            compilations.main.outputKinds('FRAMEWORK')
+        }
+    }
+    sourceSets {
+        commonMain.dependencies {
+            api project(":data:api")
+            implementation Dep.Ktor.clientCommon
+        }
+    }
 }
 
 android {
@@ -25,7 +43,6 @@ android {
 }
 
 dependencies {
-    api project(":data:api")
 
     implementation Dep.Kotlin.stdlibJvm
     implementation Dep.Kotlin.serialization
