+++ data/api-impl/src/main/java/io/github/droidkaigi/confsched2019/data/api/ApiModule.kt
@@ -15,7 +15,7 @@ import okhttp3.logging.HttpLoggingInterceptor
 @Module(includes = [ApiModule.Providers::class])
 internal abstract class ApiModule {
     @Binds abstract fun DroidKaigiApi(impl: KtorDroidKaigiApi): DroidKaigiApi
-    @Binds abstract fun GoogleFormApi(impl: KtorGoogleFormApi): GoogleFormApi
+    @Binds abstract fun GoogleFormApi(impl: InjectableKtorGoogleFormApi): GoogleFormApi
 
     @Module
     internal object Providers {
