+++ frontend/android/build.gradle
@@ -38,6 +38,14 @@ android {
         xmlOutput file("lint-results.xml")
         htmlOutput file("lint-results.html")
     }
+    packagingOptions {
+        exclude 'META-INF/ktor-http.kotlin_module'
+        exclude 'META-INF/kotlinx-io.kotlin_module'
+        exclude 'META-INF/atomicfu.kotlin_module'
+        exclude 'META-INF/ktor-utils.kotlin_module'
+        exclude 'META-INF/kotlinx-coroutines-io.kotlin_module'
+        exclude 'META-INF/ktor-client-core.kotlin_module'
+    }
 }
 
 dependencies {
